+++ client/src/api/dashboard/dashboard.api.ts
@@ -2,4 +2,4 @@ import { IDashboard } from '@/types';
 import { fetcher } from '@/utils';
 
 export const getStatisticsByWeekForUser = async (week = 0) =>
-	await fetcher<IDashboard>(`/api/v1/sleep-entries/me?week=${week}`);
+	await fetcher<IDashboard>(`/v1/sleep-entries/me?week=${week}`);