+++ client/src/api/settings/privacy.api.ts
@@ -3,10 +3,10 @@ import { IPrivacySettings } from '@/types';
 import { fetcher } from '@/utils';
 
 export const updatePrivacySettings = async (dto: SettingsPrivacyDto) =>
-	await fetcher<IPrivacySettings>('/api/v1/user-privacy-settings/me', {
+	await fetcher<IPrivacySettings>('/v1/user-privacy-settings/me', {
 		method: 'PATCH',
 		body: JSON.stringify(dto),
 	});
 
 export const getUserPrivacySettings = async () =>
-	await fetcher<IPrivacySettings>('/api/v1/user-privacy-settings/me');
+	await fetcher<IPrivacySettings>('/v1/user-privacy-settings/me');