+++ client/src/api/user/fcmToken.api.ts
@@ -2,27 +2,27 @@ import { fetcher } from '@/utils';
 
 export const saveFcmToken = async (token: string) =>
 	await fetcher(
-		'/api/v1/fcm/save-token',
+		'/v1/fcm/save-token',
 		{
 			method: 'POST',
 			body: JSON.stringify({ token }),
 		},
-		true
+		true,
 	);
 
 export const checkFcmTokenStatus = async (token: string) =>
 	await fetcher<boolean>(
-		`/api/v1/fcm/exists/${encodeURIComponent(token)}`,
+		`/v1/fcm/exists/${encodeURIComponent(token)}`,
 		{},
-		true
+		true,
 	);
 
 export const removeFcmToken = async (token: string) =>
 	await fetcher(
-		'/api/v1/fcm/remove-token',
+		'/v1/fcm/remove-token',
 		{
 			method: 'DELETE',
 			body: JSON.stringify({ token }),
 		},
-		true
+		true,
 	);