+++ client/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed TOO_MANY_REDIRECTS error after terminate session from other device.
 - Fixed Total Rest view metrics.
+- Fixed filter position while scrolling on mobile layout in shop catalog.
 
 ## [0.5.0] - 2026-04-04